+++ pycbc/pnutils.py
@@ -956,3 +956,58 @@ def hybrid_meco_frequency(m1, m2, chi1, chi2, qm1=None, qm2=None):
         qm2 = 1
 
     return velocity_to_frequency(hybrid_meco_velocity(m1, m2, chi1, chi2, qm1, qm2), m1 + m2)
+
+
+def jframe_to_l0frame(mass1, mass2, f_ref, phiref=0., thetajn=0., phijl=0.,
+                      spin1_a=0., spin2_a=0.,
+                      spin1_polar=0., spin2_polar=0.,
+                      spin12_deltaphi=0.):
+    """Converts J-frame parameters into L0 frame.
+
+    Parameters
+    ----------
+    {mass1}
+    {mass2}
+    {f_ref}
+    thetajn : float
+        Angle between the line of sight and the total angular momentume J.
+    phijl : float
+        Azimuthal angle of L on its cone about J.
+    {spin1_a}
+    {spin2_a}
+    spin1_polar : float
+        Angle between L and the spin magnitude of the larger object.
+    spin2_polar : float
+        Angle betwen L and the spin magnitude of the smaller object.
+    spin12_deltaphi : float
+        Difference between the azimuthal angles of the spin of the larger
+        object (S1) and the spin of the smaller object (S2).
+
+    Returns
+    -------
+    dict :
+        Dictionary of:
+        * {inclination}
+        * {spin1x}
+        * {spin1y}
+        * {spin1z}
+        * {spin2x}
+        * {spin2y}
+        * {spin2z}
+    """
+    inclination, spin1x, spin1y, spin1z, spin2x, spin2y, spin2z = \
+        lalsimulation.SimInspiralTransformPrecessingNewInitialConditions(
+            thetajn, phijl, spin1_polar, spin2_polar, spin12_deltaphi,
+            spin1_a, spin2_a, mass1*lal.MSUN_SI, mass2*lal.MSUN_SI, f_ref,
+            phiref)
+    out = {'inclination': inclination,
+           'spin1x': spin1x,
+           'spin1y': spin1y,
+           'spin1z': spin1z,
+           'spin2x': spin2x,
+           'spin2y': spin2y,
+           'spin2z': spin2z}
+    return out
+
+
+jframe_to_l0frame.__doc__ = _formatdocstr(jframe_to_l0frame.__doc__)
